requirements.txt
requests>
pytest>
allure-pytest>
psycopg2>